requirements3.txt
amqp==1.4.9
anyjson==0.3.3
attrs==16.1.0
billiard==3.3.0.23
celery==3.1.25
cffi==1.7.0
Cython==0.27.3
Django==1.10.6
django-appconf==1.0.2
django-celery==3.2.2
django-easy-pjax==1.3.0
django-filter==1.1.0
django-model-utils==3.0.0
django-pjax==1.4
django-redis==4.9.0
django-siteuser==0.1.2
django-users2==0.2.1
djangorestframework==3.8.2
docopt==0.6.2
enum==0.4.6
enum34==1.1.6
futures==3.2.0
grizzled-python==1.0.1
gunicorn==19.7.1
idna==2.1
ipaddress==1.0.16
kombu==3.0.37
lxml==3.7.3
multiprocessing==2.6.2.1
MySQL-python==1.2.5
olefile==0.44
pbr==2.0.0
Pillow==4.3.0
pipreqs==0.4.9
prettytable==0.7.2
protobuf==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==16.1.0
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.13.0
sae-python-dev==1.3.6
service-identity==16.0.0
six==1.10.0
socialoauth==0.3.3
sqlcmd==0.7.1
stevedore==1.21.0
uWSGI==2.0.15
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
web.py==0.38
Werkzeug==0.12.1
yarg==0.1.9
zope.interface==4.2.0
easy-thumbnails==2.5
